ref(ev_msg): split mod.rs into multiple files
