+++ src/event/event_data/delegated/mod.rs
@@ -1,8 +1,8 @@
 use super::EventData;
 use super::InceptionEvent;
-use crate::event_message::KeyEvent;
 use crate::event_message::SaidEvent;
 use crate::event_message::dummy_event::DummyInceptionEvent;
+use crate::event_message::key_event_message::KeyEvent;
 use crate::{
     derivation::self_addressing::SelfAddressing,
     error::Error,
